로그 생성 로직을 별도 모듈로 분리
로그 생성 로직을 별도 모듈로 분리하여 코드 재사용성을 높였습니다.

주요 변경사항:
1. `console-logger.ts` 파일을 생성하여 로그 생성 로직을 분리
2. `WinstonLogger.service.ts`에서 분리된 로그 생성 로직을 가져와 사용하도록 수정
3. 환경 변수에 따라 다른 로그 전송 방식을 설정하는 로직을 `console-logger.ts`로 이동
